requirements.txt
streamlit>=1.30.0
openai>=1.10.0
langchain>=0.1.12
langchain-community>=0.0.27
langchain-core>=0.1.34
python-dotenv>=1.0.1
pillow>=10.0.0
pdfplumber>=0.10.2
PyMuPDF>=1.23.10
faiss-cpu>=1.7.4
torch>=2.0.0